+++ src/main/java/nl/nn/testtool/storage/file/Storage.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2018 Nationale-Nederlanden
+   Copyright 2018 Nationale-Nederlanden, 2020 WeAreFrank!
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -126,6 +126,10 @@ public class Storage implements nl.nn.testtool.storage.LogStorage {
 
 	public void storeWithoutException(Report report) {
 		writer.storeWithoutException(report, false);
+	}
+
+	public String getWarningsAndErrors() {
+		return writer.getWarningsAndErrors();
 	}
 
 	public int getSize() throws StorageException {