+++ src/main/java/nl/nn/testtool/storage/file/TestStorage.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2018 Nationale-Nederlanden
+   Copyright 2018 Nationale-Nederlanden, 2020 WeAreFrank!
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -199,7 +199,7 @@ public class TestStorage implements nl.nn.testtool.storage.CrudStorage {
 						persistentMetadata, searchValues ,
 						MetadataExtractor.VALUE_TYPE_STRING,
 						writer[source].getMetadataFileLastModified());
-				writer[destination].store(reportBytes, (List)metadata.get(0));
+				writer[destination].store(report.getName(), reportBytes, (List)metadata.get(0));
 			}
 		}
 		writer[destination].latestStorageId = writer[source].latestStorageId;